// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Flauta.ts
export interface Flauta {
  flauta: string;
  factor1: number;
  factor2: number;
  factor3: number;
  pegamento: number;
  corrugado: string;
  laminasmt: number;
  factormts: number;
  piezaspulgadas: number;
}
export class Flauta {
  flauta: string;
  factor1: number;
  factor2: number;
  factor3: number;
  pegamento: number;
  corrugado: string;
  laminasmt: number;
  factormts: number;
  piezaspulgadas: number;
}
export interface Corrugado{
  corrugado: string;
}
export class Corrugado{
  corrugado: string;
}